:zap: Support `require 'red-amber'`, as well
